backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
anthropic
openai
google-generativeai
python-multipart
Pillow
httpx
python-jose[cryptography]
bcrypt==3.2.0
passlib[bcrypt]
certifi
gunicorn

# Database dependencies
sqlalchemy
psycopg2-binary
alembic
